Revert "Estructura MVC y vista principal agregada."
