+++ src/bancoproyecto/UsuarioControlador.java
@@ -1,19 +1,62 @@
-package bancoproyecto.view;
+package bancoproyecto;
 
-import javax.swing.*;
-import java.awt.*;
+import javax.swing.JLabel;
+import javax.swing.JOptionPane;
+import javax.swing.JPanel;
+import javax.swing.JPasswordField;
+import javax.swing.JTextField;
+import java.awt.GridLayout;
 
-public class CredencialesUtils {
+public class UsuarioControlador {
+    /**
+     * Crea un usuario con los datos ingresados por el usuario.
+     * 
+     * @return El usuario creado.
+     */
+    public static Usuario CreaUsuario() {
+        String usuario = SolicitaUsuario(false);
+        String contrasena = SolicitaContrasena(false);
+        return new Usuario(usuario, contrasena);
+    }
+
+    /**
+     * Inicia sesion con los datos ingresados por el usuario.
+     */
+    public static void IniciarSesionUsuario() {
+        // Solicitar usuario y contraseña
+        String usuario = SolicitaUsuario(true);
+        String contrasena = SolicitaContrasena(true);
+
+        // Crear un objeto de tipo Usuario con los datos ingresados por el usuario.
+        Usuario usuarioInicioSesion = new Usuario(usuario, contrasena);
+
+        // Validar las credenciales del usuario.
+        Boolean inicioSesionResultado = UsuarioData.InicioSesion(usuarioInicioSesion);
+
+        if (inicioSesionResultado) {
+            JOptionPane.showMessageDialog(null, "Inicio de sesion exitoso", "Inicio de sesion",
+                    JOptionPane.INFORMATION_MESSAGE);
+
+            // TODO: Agregar el código para mostrar el menu principal.
+
+            System.exit(0); // ¡Cierre de proyecto exitoso!
+
+        } else {
+            JOptionPane.showMessageDialog(null, "Inicio de sesion fallido", "Inicio de sesion",
+                    JOptionPane.ERROR_MESSAGE);
+            IniciarSesionUsuario(); // Volver a solicitar usuario y contraseña.
+        }
+    }
 
     /**
      * Solicita el usuario al usuario.
-     *
+     * 
      * @param esInicioSesion Indica si el usuario se solicita para iniciar sesion o
      *                       para crear un usuario.
-     *
+     * 
      * @return El nombre de usuario ingresado por el usuario.
      */
-    public static String SolicitaUsuario(Boolean esInicioSesion) {
+    private static String SolicitaUsuario(Boolean esInicioSesion) {
         String usuario = null; // Inicializar la variable para que no marque error.
         boolean validInput = false; // Bandera para validar la entrada del usuario.
 
@@ -38,20 +81,15 @@ public class CredencialesUtils {
                 if (opcion == JOptionPane.OK_OPTION) {
                     usuario = userField.getText();
                 } else {
-                    validInput = true;
-                    return null;
+                    var salir = JOptionPane.showConfirmDialog(null, "¿Desear salir?", "Salir",
+                            JOptionPane.YES_NO_OPTION);
+
+                    if (salir == JOptionPane.YES_OPTION) {
+                        System.exit(0);
+                    }
+
+                    throw new Exception("Ingrese de nuevo el usuario");
                 }
-//                else
-//                {
-//                    var salir = JOptionPane.showConfirmDialog(null, "¿Desear salir?", "Salir",
-//                            JOptionPane.YES_NO_OPTION);
-//
-//                    if (salir == JOptionPane.YES_OPTION) {
-//                        System.exit(0);
-//                    }
-//
-//                    throw new Exception("Ingrese de nuevo el usuario");
-//                }
 
                 if (usuario.isEmpty()) {
                     throw new Exception("El usuario no puede estar vació");
@@ -79,12 +117,12 @@ public class CredencialesUtils {
 
     /**
      * Solicita la contrasena al usuario.
-     *
+     * 
      * @param esInicioSesion Indica si la contraseña se solicita para iniciar sesión
      *                       o para crear un usuario.
      * @return La contraseña ingresada por el usuario.
      */
-    public static String SolicitaContrasena(Boolean esInicioSesion) {
+    private static String SolicitaContrasena(Boolean esInicioSesion) {
         String contrasena = null; // Inicializar la variable para que no marque error.
         boolean validInput = false; // Bandera para validar la entrada del usuario.
 
@@ -110,20 +148,15 @@ public class CredencialesUtils {
                 if (opcion == JOptionPane.OK_OPTION) {
                     contrasena = new String(passwordField.getPassword());
                 } else {
-                    validInput = true;
-                    return null;
+                    var salir = JOptionPane.showConfirmDialog(null, "¿Desear salir?", "Salir",
+                            JOptionPane.YES_NO_OPTION);
+
+                    if (salir == JOptionPane.YES_OPTION) {
+                        System.exit(0);
+                    }
+
+                    throw new Exception("Ingrese de nuevo la contrasena");
                 }
-//                else
-//                {
-//                    var salir = JOptionPane.showConfirmDialog(null, "¿Desear salir?", "Salir",
-//                            JOptionPane.YES_NO_OPTION);
-//
-//                    if (salir == JOptionPane.YES_OPTION) {
-//                        System.exit(0);
-//                    }
-//
-//                    throw new Exception("Ingrese de nuevo la contrasena");
-//                }
 
                 // #region Validaciones de contraseña
                 if (contrasena.isEmpty()) {
